+++ src/components/datagrid/data_grid.tsx
@@ -55,6 +55,7 @@ import {
   EuiDataGridToolBarVisibilityOptions,
   EuiDataGridFocusedCell,
   EuiDataGridOnColumnResizeHandler,
+  EuiDataGridStyleFooter,
 } from './data_grid_types';
 import { EuiDataGridCellProps } from './data_grid_cell';
 import { EuiButtonEmpty } from '../button';
@@ -120,6 +121,12 @@ type CommonGridProps = CommonProps &
      */
     renderCellValue: EuiDataGridCellProps['renderCellValue'];
     /**
+     * A function called to render a cell's value. Behind the scenes it is treated as a React component
+     * allowing hooks, context, and other React concepts to be used. The function receives a #CellValueElement
+     * as its only argument.
+     */
+    renderFooterCellValue?: EuiDataGridCellProps['renderCellValue'];
+    /**
      * Defines the look and feel for the grid. Accepts a partial #EuiDataGridStyle object. Settings provided may be overwritten or merged with user defined preferences if toolbarVisibility density controls are available.
      */
     gridStyle?: EuiDataGridStyle;
@@ -165,6 +172,12 @@ const fontSizesToClassMap: { [size in EuiDataGridStyleFontSizes]: string } = {
 const headerToClassMap: { [header in EuiDataGridStyleHeader]: string } = {
   shade: 'euiDataGrid--headerShade',
   underline: 'euiDataGrid--headerUnderline',
+};
+
+const footerToClassMap: { [footer in EuiDataGridStyleFooter]: string } = {
+  shade: 'euiDataGrid--footerShade',
+  overline: 'euiDataGrid--footerOverline',
+  striped: '',
 };
 
 const rowHoverToClassMap: {
@@ -462,7 +475,7 @@ function createKeyDownHandler(
 
     if (key === keys.ARROW_DOWN) {
       event.preventDefault();
-      if (y < rowCount - 1) {
+      if (props.renderFooterCellValue ? y < rowCount : y < rowCount - 1) {
         setFocusedCell([x, y + 1]);
       }
     } else if (key === keys.ARROW_LEFT) {
@@ -682,6 +695,7 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = props => {
     schemaDetectors,
     rowCount,
     renderCellValue,
+    renderFooterCellValue,
     className,
     gridStyle,
     toolbarVisibility = true,
@@ -799,10 +813,14 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = props => {
     fontSizesToClassMap[gridStyles.fontSize!],
     bordersToClassMap[gridStyles.border!],
     headerToClassMap[gridStyles.header!],
+    footerToClassMap[gridStyles.footer!],
     rowHoverToClassMap[gridStyles.rowHover!],
     cellPaddingsToClassMap[gridStyles.cellPadding!],
     {
       'euiDataGrid--stripes': gridStyles.stripes!,
+    },
+    {
+      'euiDataGrid--stickyFooter': gridStyles.footer && gridStyles.stickyFooter,
     },
     {
       'euiDataGrid--fullScreen': isFullScreen,
@@ -909,6 +927,18 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = props => {
   const gridIds = htmlIdGenerator();
   const gridId = gridIds();
   const ariaLabelledById = gridIds();
+
+  const commonGridProps = {
+    columns: orderedVisibleColumns,
+    columnWidths,
+    defaultColumnWidth,
+    focusedCell,
+    leadingControlColumns,
+    onCellFocus: setFocusedCell,
+    schema: mergedSchema,
+    sorting,
+    trailingControlColumns,
+  };
 
   return (
     <EuiI18n
@@ -1024,48 +1054,25 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = props => {
                                       {ref => (
                                         <EuiDataGridHeaderRow
                                           ref={ref}
-                                          leadingControlColumns={
-                                            leadingControlColumns
-                                          }
-                                          trailingControlColumns={
-                                            trailingControlColumns
-                                          }
-                                          columns={orderedVisibleColumns}
-                                          columnWidths={columnWidths}
-                                          defaultColumnWidth={
-                                            defaultColumnWidth
-                                          }
+                                          {...commonGridProps}
                                           setColumnWidth={setColumnWidth}
-                                          schema={mergedSchema}
-                                          sorting={sorting}
                                           headerIsInteractive={
                                             headerIsInteractive
                                           }
-                                          focusedCell={focusedCell}
-                                          setFocusedCell={setFocusedCell}
                                         />
                                       )}
                                     </EuiMutationObserver>
                                     <EuiDataGridBody
-                                      columnWidths={columnWidths}
-                                      defaultColumnWidth={defaultColumnWidth}
+                                      {...commonGridProps}
                                       inMemoryValues={inMemoryValues}
                                       inMemory={inMemory}
-                                      leadingControlColumns={
-                                        leadingControlColumns
-                                      }
-                                      trailingControlColumns={
-                                        trailingControlColumns
-                                      }
-                                      columns={orderedVisibleColumns}
-                                      schema={mergedSchema}
                                       schemaDetectors={allSchemaDetectors}
                                       popoverContents={popoverContents}
-                                      focusedCell={focusedCell}
-                                      onCellFocus={setFocusedCell}
                                       pagination={pagination}
-                                      sorting={sorting}
                                       renderCellValue={renderCellValue}
+                                      renderFooterCellValue={
+                                        renderFooterCellValue
+                                      }
                                       rowCount={rowCount}
                                       interactiveCellId={interactiveCellId}
                                     />